health, imu, log, and test update with new uint format

+++ src/application/health_checks/health_checks.c
@@ -32,7 +32,7 @@
 typedef struct {
 	TaskHandle_t task_handle;
 	bool is_kicked;
-	float last_kick_timestamp;
+	uint32_t last_kick_timestamp; // ms
 	uint32_t timeout_ticks;
 } watchdog_task_t;
 
@@ -59,7 +59,7 @@ uint32_t check_current(void) {
 	uint32_t adc_current_mA;
 
 	if (get_adc_current(&adc_current_mA) == W_SUCCESS) {
-		float ms = 0;
+		uint32_t ms = 0;
 		timer_get_ms(&ms);
 		can_msg_t msg = {0};
 
@@ -85,7 +85,7 @@ w_status_t health_check_init(void) {
 
 w_status_t watchdog_kick(void) {
 	TaskHandle_t current_task = xTaskGetCurrentTaskHandle();
-	float current_time = 0.0f;
+	uint32_t current_time = 0;
 	w_status_t status = W_SUCCESS;
 
 	status |= timer_get_ms(&current_time);
@@ -124,7 +124,7 @@ w_status_t watchdog_register_task(TaskHandle_t task_handle, uint32_t timeout_tic
 		}
 	}
 
-	float current_time = 0.0f;
+	uint32_t current_time = 0;
 	w_status_t status = W_SUCCESS;
 	status |= timer_get_ms(&current_time);
 
@@ -149,7 +149,7 @@ w_status_t watchdog_register_task(TaskHandle_t task_handle, uint32_t timeout_tic
  */
 uint32_t check_watchdog_tasks(void) {
 	uint32_t status_bitfield = 0;
-	float current_time = 0.0f;
+	uint32_t current_time = 0;
 
 	// failing to get time isn't a fatal err
 	if (timer_get_ms(&current_time) != W_SUCCESS) {
@@ -158,7 +158,7 @@ uint32_t check_watchdog_tasks(void) {
 	}
 
 	for (uint32_t i = 0; i < num_watchdog_tasks; i++) {
-		float time_elapsed = current_time - watchdog_tasks[i].last_kick_timestamp;
+		uint32_t time_elapsed = current_time - watchdog_tasks[i].last_kick_timestamp;
 		uint32_t ticks_elapsed = pdMS_TO_TICKS((uint32_t)time_elapsed); // time to ticks
 
 		if (watchdog_tasks[i].is_kicked || (ticks_elapsed <= watchdog_tasks[i].timeout_ticks)) {
